+++ interfaces/python/exago_python.cpp
@@ -1,6 +1,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 #include <mpi.h>
+#include <mpi4py/mpi4py.h>
 #include <pflow.h>
 #include <private/pflowimpl.h>
 #include <opflow.h>
@@ -33,7 +34,14 @@ constexpr int num_gen_bus_voltage_types = 0
 
 static inline auto prefix() -> const char * { return CMAKE_INSTALL_PREFIX; }
 
-int initialize(char *appname) {
+int initialize(char *appname, MPI_Comm comm) {
+  PetscErrorCode ierr;
+  ierr = ExaGOInitialize(comm, nullptr, nullptr, appname, nullptr);
+  ExaGOCheckError(ierr);
+  return ierr;
+}
+
+int initialize_no_comm(char *appname) {
   PetscErrorCode ierr;
   MPI_Comm communicator;
   ierr = ExaGOGetSelfCommunicator(&communicator);
@@ -43,12 +51,34 @@ int initialize(char *appname) {
   return ierr;
 }
 
+/*! Return a MPI communicator from mpi4py communicator object. */
+MPI_Comm *get_mpi_comm(pybind11::object py_comm) {
+  auto comm_ptr = PyMPIComm_Get(py_comm.ptr());
+
+  if (!comm_ptr)
+    throw pybind11::error_already_set();
+
+  return comm_ptr;
+}
+
 PYBIND11_MODULE(exago, m) {
+  // Initialize mpi4py's C-API
+  if (import_mpi4py() < 0) {
+    // mpi4py calls the Python C API
+    // we let pybind11 give us the detailed traceback
+    throw pybind11::error_already_set();
+  }
+
   m.doc() = "Python wrapper for ExaGO.";
 
   /* Bindings for top-level utilities, such as initialization, finalization,
    * and helpers for interacting with enums and macros */
-  m.def("initialize", &initialize);
+  // Initialize with pybind11 comm
+  m.def("initialize", [](char *appname, pybind11::object py_comm) {
+    auto comm = get_mpi_comm(py_comm);
+    return initialize(appname, *comm);
+  });
+  m.def("initialize", &initialize_no_comm);
   m.def("finalize", &ExaGOFinalize);
   m.def("prefix", &prefix);